backend/requirements.txt
# OCR Marks Card Extractor Requirements
paddlepaddle>=3.2.0
paddlex>=3.0.0
paddleocr>=3.2.0
pillow>=10.0.0
opencv-contrib-python>=4.10.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Flask Backend Requirements
Flask>=3.0.0
Werkzeug>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
Flask-Mail>=0.9.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
